Abstracting Renderer into class and utilization macros and rearranging stuff
