+++ OpenGL/src/Application.cpp
@@ -5,6 +5,7 @@
 #include <sstream>
 #include <GLFW/glfw3.h>
 
+#include "VertexBufferLayout.h"
 #include "Renderer.h"
 #include "IndexBuffer.h"
 #include "VertexArray.h"
@@ -75,19 +76,20 @@ int main(void)
         vb.Unbind();
         ib.Unbind();
 
+        Renderer renderer;
+
         float r = 0.0f;
         float increment = 0.05f;
         /* Loop until the user closes the window */
         while (!glfwWindowShouldClose(window))
         {
             /* Render here */
-            GLCall(glClear(GL_COLOR_BUFFER_BIT));
+            renderer.Clear();
 
             shader.Bind();
             shader.SetUniform4f("u_Color", r, 0.3f, 0.8f, 1.0);
 
-            va.Bind();
-            ib.Bind();
+            renderer.Draw(va, ib, shader);
 
             GLCall(glDrawElements(GL_TRIANGLES, 6, GL_UNSIGNED_INT, nullptr));
