+++ OpenGL/src/Renderer.cpp
@@ -1,19 +1,17 @@
 #include "Renderer.h"
+#include "UtilFunctions.h"
 
 #include <iostream>
 
-void GLClearError()
+void Renderer::Clear() const
 {
-    while (glGetError() != GL_NO_ERROR);
+    GLCall(glClear(GL_COLOR_BUFFER_BIT));
 }
 
-bool GLLogCall(const char* function, const char* file, int line)
+void Renderer::Draw(const VertexArray& va, const IndexBuffer& ib, const Shader& shader) const
 {
-    while (GLenum error = glGetError())
-    {
-        std::cout << "[OpenGL Error] (" << error << ")" << function
-            << " " << file << ":" << line << std::endl;
-        return false;
-    }
-    return true;
+    shader.Bind();
+    va.Bind();
+    ib.Bind();
+    GLCall(glDrawElements(GL_TRIANGLES, ib.GetCount(), GL_UNSIGNED_INT, nullptr));
 }
