+++ OpenGL/src/Renderer.h
@@ -2,10 +2,13 @@
 
 #include <GL/glew.h>
 
-#define ASSERT(x) if (!(x)) __debugbreak();
-#define GLCall(x) GLClearError();\
-    x;\
-    ASSERT(GLLogCall(#x, __FILE__, __LINE__)) // # before x turns the x into a string, __FILE__ and __LINE__ are for all compilers
+#include "VertexArray.h"
+#include "IndexBuffer.h"
+#include "Shader.h"
 
-void GLClearError();
-bool GLLogCall(const char* function, const char* file, int line);
+class Renderer
+{
+public:
+    void Clear() const;
+    void Draw(const VertexArray& va, const IndexBuffer& ib, const Shader& shader) const;
+};
