+++ OpenGL/src/UtilFunctions.cpp
@@ -1,0 +1,19 @@
+#include <iostream>
+#include <GL/glew.h>
+#include "UtilFunctions.h"
+
+void GLClearError()
+{
+    while (glGetError() != GL_NO_ERROR);
+}
+
+bool GLLogCall(const char* function, const char* file, int line)
+{
+    while (GLenum error = glGetError())
+    {
+        std::cout << "[OpenGL Error] (" << error << ")" << function
+            << " " << file << ":" << line << std::endl;
+        return false;
+    }
+    return true;
+}
